添加根据跟班id查询用户信息feign1.3

+++ game-base/src/main/java/com/dkm/feign/ResourceFeignClient.java
@@ -93,7 +93,7 @@ public interface ResourceFeignClient {
      * @return
      */
     @GetMapping("/queryAidUser")
-    Result<AttendantUserVo> queryAidUser (@RequestParam(value = "CaughtPeopleId")  Integer CaughtPeopleId);
+    Result<AttendantUserVo> queryAidUser (@RequestParam(value = "CaughtPeopleId")  Long CaughtPeopleId);
 
 
 }
